refactor: move code a bit and clean it up
