+++ src/rnapolis/tertiary.py
@@ -1,14 +1,22 @@
 import logging
 import math
+from collections import defaultdict
 from dataclasses import dataclass, field
-from typing import IO, Dict, List, Optional, Tuple, Union
+from typing import Dict, List, Optional, Set, Tuple, Union
 
 import numpy
 import numpy.typing
-from mmcif.io.IoAdapterPy import IoAdapterPy
 
-from rnapolis.common import GlycosidicBond, LeontisWesthof, ResidueAuth, ResidueLabel
-from rnapolis.secondary import BasePair, Residue, Stacking
+from rnapolis.common import (
+    BasePair,
+    GlycosidicBond,
+    LeontisWesthof,
+    Residue,
+    ResidueAuth,
+    ResidueLabel,
+    Stacking,
+    Structure2D,
+)
 
 BASE_ATOMS = {
     "A": ["N1", "C2", "N3", "C4", "C5", "C6", "N6", "N7", "C8", "N9"],
@@ -83,7 +91,7 @@ BASE_EDGES = {
 }
 
 
-@dataclass(frozen=True)
+@dataclass(frozen=True, order=True)
 class Atom:
     label: Optional[ResidueLabel]
     auth: Optional[ResidueAuth]
@@ -98,9 +106,8 @@ class Atom:
         return numpy.array([self.x, self.y, self.z])
 
 
-@dataclass(order=True)
+@dataclass(frozen=True, order=True)
 class Residue3D(Residue):
-    index: int
     model: int
     one_letter_name: str
     atoms: Tuple[Atom, ...]
@@ -186,7 +193,7 @@ class Residue3D(Residue):
             self.find_atom("C4"),
         ]
         if all([atom is not None for atom in atoms]):
-            return Residue3D.__torsion_angle(atoms)  # type: ignore
+            return torsion_angle(*atoms)  # type: ignore
         return math.nan
 
     def __chi_pyrimidine(self) -> float:
@@ -197,7 +204,7 @@ class Residue3D(Residue):
             self.find_atom("C2"),
         ]
         if all([atom is not None for atom in atoms]):
-            return Residue3D.__torsion_angle(atoms)  # type: ignore
+            return torsion_angle(*atoms)  # type: ignore
         return math.nan
 
     def __outer_generator(self):
@@ -252,7 +259,7 @@ class Residue3D(Residue):
         yield Atom(self.label, self.auth, self.model, "UNK", 0.0, 0.0, 0.0)
 
 
-@dataclass
+@dataclass(frozen=True, order=True)
 class BasePair3D(BasePair):
     nt1_3d: Residue3D
     nt2_3d: Residue3D
@@ -310,31 +317,19 @@ class BasePair3D(BasePair):
             nts == "AU" or nts == "AT" or nts == "CG" or nts == "GU"
         )
 
-    def conflicts_with(self, other) -> bool:
-        xi, yi = sorted([self.nt1_3d.index, self.nt2_3d.index])
-        xj, yj = sorted([other.nt1_3d.index, other.nt2_3d.index])
-        return xi < xj < yi < yj or xj < xi < yj < yi
 
-    def in_tetrad(self, analysis) -> bool:
-        for tetrad in analysis.tetrads:
-            if self in (
-                tetrad.pair_12,
-                tetrad.pair_23,
-                tetrad.pair_34,
-                tetrad.pair_41,
-                tetrad.pair_12.reverse(),
-                tetrad.pair_23.reverse(),
-                tetrad.pair_34.reverse(),
-                tetrad.pair_41.reverse(),
-            ):
-                return True
-        return False
-
-
-@dataclass
+@dataclass(frozen=True, order=True)
 class Stacking3D(Stacking):
     nt1_3d: Residue3D
     nt2_3d: Residue3D
+
+    @property
+    def reverse(self):
+        if self.topology is None:
+            return self
+        return Stacking3D(
+            self.nt2, self.nt1, self.topology.reverse, self.nt2_3d, self.nt1_3d
+        )
 
 
 @dataclass
@@ -360,300 +355,86 @@ class Structure3D:
         return None
 
 
-def read_3d_structure(cif_or_pdb: IO[str], model: int) -> Structure3D:
-    atoms, modified, sequence = (
-        parse_cif(cif_or_pdb) if is_cif(cif_or_pdb) else parse_pdb(cif_or_pdb)
-    )
-    atoms = list(filter(lambda atom: atom.model == model, atoms))
-    return group_atoms(atoms, modified, sequence)
+@dataclass
+class Mapping2D3D:
+    structure3d: Structure3D
+    structure2d: Structure2D
 
-
-def is_cif(cif_or_pdb: IO[str]) -> bool:
-    cif_or_pdb.seek(0)
-    for line in cif_or_pdb.readlines():
-        if line.startswith("_atom_site"):
-            return True
-    return False
-
-
-def parse_cif(
-    cif: IO[str],
-) -> Tuple[
-    List[Atom],
-    Dict[Union[ResidueLabel, ResidueAuth], str],
-    Dict[Tuple[str, int], str],
-]:
-    cif.seek(0)
-
-    io_adapter = IoAdapterPy()
-    data = io_adapter.readFile(cif.name)
-    atoms: List[Atom] = []
-    modified: Dict[Union[ResidueLabel, ResidueAuth], str] = {}
-    sequence = {}
-
-    if data:
-        atom_site = data[0].getObj("atom_site")
-        mod_residue = data[0].getObj("pdbx_struct_mod_residue")
-        entity_poly = data[0].getObj("entity_poly")
-
-        if atom_site:
-            for row in atom_site.getRowList():
-                row_dict = dict(zip(atom_site.getAttributeList(), row))
-
-                label_chain_name = row_dict.get("label_asym_id", None)
-                label_residue_number = try_parse_int(row_dict.get("label_seq_id", None))
-                label_residue_name = row_dict.get("label_comp_id", None)
-                auth_chain_name = row_dict.get("auth_asym_id", None)
-                auth_residue_number = try_parse_int(row_dict.get("auth_seq_id", None))
-                auth_residue_name = row_dict.get("auth_comp_id", None)
-                insertion_code = row_dict.get("pdbx_PDB_ins_code", None)
-
-                # mmCIF marks empty values with ?
-                if insertion_code == "?":
-                    insertion_code = None
-
-                if label_chain_name is None and auth_chain_name is None:
-                    raise RuntimeError(
-                        f"Cannot parse an atom line with empty chain name: {row}"
-                    )
-                if label_residue_number is None and auth_residue_number is None:
-                    raise RuntimeError(
-                        f"Cannot parse an atom line with empty residue number: {row}"
-                    )
-                if label_residue_name is None and auth_residue_name is None:
-                    raise RuntimeError(
-                        f"Cannot parse an atom line with empty residue name: {row}"
-                    )
-
-                label = None
-                if label_chain_name and label_residue_number and label_residue_name:
-                    label = ResidueLabel(
-                        label_chain_name, label_residue_number, label_residue_name
-                    )
-
-                auth = None
-                if auth_chain_name and auth_residue_number and auth_residue_name:
-                    auth = ResidueAuth(
-                        auth_chain_name,
-                        auth_residue_number,
-                        insertion_code,
-                        auth_residue_name,
-                    )
-
-                model = int(row_dict.get("pdbx_PDB_model_num", "1"))
-                atom_name = row_dict["label_atom_id"]
-                x = float(row_dict["Cartn_x"])
-                y = float(row_dict["Cartn_y"])
-                z = float(row_dict["Cartn_z"])
-                atoms.append(Atom(label, auth, model, atom_name, x, y, z))
-
-        if mod_residue:
-            for row in mod_residue.getRowList():
-                row_dict = dict(zip(mod_residue.getAttributeList(), row))
-
-                label_chain_name = row_dict.get("label_asym_id", None)
-                label_residue_number = try_parse_int(row_dict.get("label_seq_id", None))
-                label_residue_name = row_dict.get("label_comp_id", None)
-                auth_chain_name = row_dict.get("auth_asym_id", None)
-                auth_residue_number = try_parse_int(row_dict.get("auth_seq_id", None))
-                auth_residue_name = row_dict.get("auth_comp_id", None)
-                insertion_code = row_dict.get("PDB_ins_code", None)
-
-                label = None
-                if label_chain_name and label_residue_number and label_residue_name:
-                    label = ResidueLabel(
-                        label_chain_name, label_residue_number, label_residue_name
-                    )
-
-                auth = None
-                if (
-                    auth_chain_name
-                    and auth_residue_number
-                    and auth_residue_name
-                    and insertion_code
-                ):
-                    auth = ResidueAuth(
-                        auth_chain_name,
-                        auth_residue_number,
-                        insertion_code,
-                        auth_residue_name,
-                    )
-
-                # TODO: is processing this data for each model separately required?
-                # model = row_dict.get('PDB_model_num', '1')
-                standard_residue_name = row_dict.get("parent_comp_id", "n")
-
-                if label is not None:
-                    modified[label] = standard_residue_name
-                if auth is not None:
-                    modified[auth] = standard_residue_name
-
-        if entity_poly:
-            for row in entity_poly.getRowList():
-                row_dict = dict(zip(entity_poly.getAttributeList(), row))
-
-                pdbx_strand_id = row_dict.get("pdbx_strand_id", None)
-                pdbx_seq_one_letter_code_can = row_dict.get(
-                    "pdbx_seq_one_letter_code_can", None
+    @property
+    def base_pairs(self) -> List[BasePair3D]:
+        result = []
+        used = set()
+        for base_pair in self.structure2d.basePairs:
+            nt1 = self.structure3d.find_residue(base_pair.nt1.label, base_pair.nt1.auth)
+            nt2 = self.structure3d.find_residue(base_pair.nt2.label, base_pair.nt2.auth)
+            if nt1 is not None and nt2 is not None:
+                bp = BasePair3D(
+                    base_pair.nt1,
+                    base_pair.nt2,
+                    base_pair.lw,
+                    base_pair.saenger,
+                    nt1,
+                    nt2,
                 )
+                if bp not in used:
+                    result.append(bp)
+                    used.add(bp)
+                if bp.reverse not in used:
+                    result.append(bp.reverse)
+                    used.add(bp.reverse)
+        return result
 
-                if pdbx_strand_id and pdbx_seq_one_letter_code_can:
-                    for strand in pdbx_strand_id.split(","):
-                        for i, letter in enumerate(pdbx_seq_one_letter_code_can):
-                            sequence[(strand, i + 1)] = letter
+    @property
+    def base_pair_graph(
+        self,
+    ) -> Dict[Residue3D, Set[Residue3D]]:
+        graph = defaultdict(set)
+        for pair in self.base_pairs:
+            graph[pair.nt1_3d].add(pair.nt2_3d)
+            graph[pair.nt2_3d].add(pair.nt1_3d)
+        return graph
 
-    return atoms, modified, sequence
+    @property
+    def base_pair_dict(self) -> Dict[Tuple[Residue3D, Residue3D], BasePair3D]:
+        result = {}
+        for base_pair in self.base_pairs:
+            residue_i = base_pair.nt1_3d
+            residue_j = base_pair.nt2_3d
+            result[(residue_i, residue_j)] = base_pair
+            result[(residue_j, residue_i)] = base_pair.reverse
+        return result
 
+    @property
+    def stackings(self) -> List[Stacking3D]:
+        result = []
+        used = set()
+        for stacking in self.structure2d.stackings:
+            nt1 = self.structure3d.find_residue(stacking.nt1.label, stacking.nt1.auth)
+            nt2 = self.structure3d.find_residue(stacking.nt2.label, stacking.nt2.auth)
+            if nt1 is not None and nt2 is not None:
+                st = Stacking3D(stacking.nt1, stacking.nt2, stacking.topology, nt1, nt2)
+                if st not in used:
+                    result.append(st)
+                    used.add(st)
+                if st.reverse not in used:
+                    result.append(st.reverse)
+                    used.add(st.reverse)
+        return result
 
-def parse_pdb(
-    pdb: IO[str],
-) -> Tuple[
-    List[Atom],
-    Dict[Union[ResidueLabel, ResidueAuth], str],
-    Dict[Tuple[str, int], str],
-]:
-    pdb.seek(0)
-    atoms: List[Atom] = []
-    modified: Dict[Union[ResidueLabel, ResidueAuth], str] = {}
-    model = 1
-
-    for line in pdb.readlines():
-        if line.startswith("MODEL"):
-            model = int(line[10:14].strip())
-        elif line.startswith("ATOM") or line.startswith("HETATM"):
-            alternate_location = line[16]
-            if alternate_location != " ":
-                continue
-            atom_name = line[12:16].strip()
-            residue_name = line[18:20].strip()
-            chain_identifier = line[21]
-            residue_number = int(line[22:26].strip())
-            insertion_code = line[26] if line[26] != " " else None
-            x = float(line[30:38].strip())
-            y = float(line[38:46].strip())
-            z = float(line[46:54].strip())
-            auth = ResidueAuth(
-                chain_identifier, residue_number, insertion_code, residue_name
-            )
-            atoms.append(Atom(None, auth, model, atom_name, x, y, z))
-        elif line.startswith("MODRES"):
-            original_name = line[12:15]
-            chain_identifier = line[16]
-            residue_number = int(line[18:22].strip())
-            insertion_code = line[23]
-            standard_residue_name = line[24:27].strip()
-            auth = ResidueAuth(
-                chain_identifier, residue_number, insertion_code, original_name
-            )
-            modified[auth] = standard_residue_name
-
-    return atoms, modified, {}
-
-
-def group_atoms(
-    atoms: List[Atom],
-    modified: Dict[Union[ResidueLabel, ResidueAuth], str],
-    sequence: Dict[Tuple[str, int], str],
-) -> Structure3D:
-    if not atoms:
-        return Structure3D([])
-
-    key_previous = (atoms[0].label, atoms[0].auth, atoms[0].model)
-    residue_atoms = [atoms[0]]
-    residues: List[Residue3D] = []
-    index = 1
-
-    for atom in atoms[1:]:
-        key = (atom.label, atom.auth, atom.model)
-        if key == key_previous:
-            residue_atoms.append(atom)
-        else:
-            label = key_previous[0]
-            auth = key_previous[1]
-            model = key_previous[2]
-            name = get_residue_name(auth, label, modified)
-            one_letter_name = get_one_letter_name(label, sequence, name)
-            if one_letter_name not in "ACGUT":
-                one_letter_name = detect_one_letter_name(residue_atoms)
-            residues.append(
-                Residue3D(
-                    label, auth, index, model, one_letter_name, tuple(residue_atoms)
-                )
-            )
-            index += 1
-            key_previous = key
-            residue_atoms = [atom]
-
-    label = key_previous[0]
-    auth = key_previous[1]
-    model = key_previous[2]
-    name = get_residue_name(auth, label, modified)
-    one_letter_name = get_one_letter_name(label, sequence, name)
-    if one_letter_name not in "ACGUT":
-        one_letter_name = detect_one_letter_name(residue_atoms)
-    residues.append(
-        Residue3D(label, auth, index, model, one_letter_name, tuple(residue_atoms))
-    )
-
-    return Structure3D(residues)
+    @property
+    def stacking_graph(self) -> Dict[Residue3D, Set[Residue3D]]:
+        graph = defaultdict(set)
+        for pair in self.stackings:
+            graph[pair.nt1_3d].add(pair.nt2_3d)
+            graph[pair.nt2_3d].add(pair.nt1_3d)
+        return graph
 
 
-def get_residue_name(
-    auth: Optional[ResidueAuth],
-    label: Optional[ResidueLabel],
-    modified: Dict[Union[ResidueAuth, ResidueLabel], str],
-) -> str:
-    if auth is not None and auth in modified:
-        name = modified[auth].lower()
-    elif label is not None and label in modified:
-        name = modified[label].lower()
-    elif auth is not None:
-        name = auth.name
-    elif label is not None:
-        name = label.name
-    else:
-        # any nucleotide
-        name = "n"
-    return name
-
-
-def get_one_letter_name(
-    label: Optional[ResidueLabel], sequence: Dict[Tuple[str, int], str], name: str
-) -> str:
-    # try getting the value from _entity_poly first
-    if label is not None:
-        key = (label.chain, label.number)
-        if key in sequence:
-            return sequence[key]
-
-    # RNA
-    if len(name) == 1:
-        return name
-    # DNA
-    if len(name) == 2 and name[0].upper() == "D":
-        return name[1]
-    # try the last letter of the name
-    if str.isalpha(name[-1]):
-        return name[-1]
-    # any nucleotide
-    return "n"
-
-
-def detect_one_letter_name(atoms: List[Atom]) -> str:
-    atom_names_present = {atom.name for atom in atoms}
-    score = {}
-    for candidate in "ACGUT":
-        atom_names_expected = BASE_ATOMS[candidate]
-        count = sum(
-            1 for atom in atom_names_expected if atom in atom_names_present
-        ) / len(atom_names_expected)
-        score[candidate] = count
-    items = sorted(score.items(), key=lambda kv: kv[1], reverse=True)
-    return items[0][0]
-
-
-def try_parse_int(s: str) -> Optional[int]:
-    try:
-        return int(s)
-    except:
-        return None
+def torsion_angle(a1: Atom, a2: Atom, a3: Atom, a4: Atom) -> float:
+    v1 = a2.coordinates - a1.coordinates
+    v2 = a3.coordinates - a2.coordinates
+    v3 = a4.coordinates - a3.coordinates
+    t1: numpy.typing.NDArray[numpy.floating] = numpy.cross(v1, v2)
+    t2: numpy.typing.NDArray[numpy.floating] = numpy.cross(v2, v3)
+    t3: numpy.typing.NDArray[numpy.floating] = v1 * numpy.linalg.norm(v2)
+    return math.atan2(numpy.dot(t2, t3), numpy.dot(t1, t2))
